افزودن فیچر گزارشات پرونده

+++ CompanyManagment.App.Contracts/FileState/FileStateViewModel.cs
@@ -6,7 +6,7 @@ namespace CompanyManagment.App.Contracts.FileState
     {
         public long Id { get; set; }
         public int State { get; set; }
-        public int FileTiming_Id { get; set; }
+        public long FileTiming_Id { get; set; }
         public string Title { get; set; }
     }
 }
